server/projects/file/requirements.txt
# File Server Requirements
# for server deploy
gunicorn==20.1.0

# for django
Django==3.1.12

# for django using mysql server
pymysql==1.0.2

# for rest api
djangorestframework==3.11.1

# for celery
celery==4.3.0
vine==1.3.0
redis==3.3.11

# for exception log
sentry-sdk==0.10.2

# for storage
cos-python-sdk-v5==1.7.4
minio==7.1.0

# for password
pyaes==1.6.1
pycryptodome==3.9.8
cryptography==3.2.1

# for monitor
django-prometheus==2.0.0
